backend/requirements.txt
fastapi
uvicorn
sqlmodel
pypdf
beautifulsoup4
groq
python-dotenv
python-multipart
requests
